clinks-Api-main/requirements.txt
amqp==2.6.1
apns2==0.7.2
asgiref==3.3.4
billiard==3.6.4.0
boto==2.49.0
Brotli==1.0.9
celery==4.4.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
cssselect2==0.6.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.11.0
-e git+https://github.com/bertonha/django-push-notifications.git@7cb6936e93b3a67332ff72422273b02e8dc37691#egg=django_push_notifications
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
fonttools==4.33.2
freezegun==1.1.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
h2==2.6.2
haversine==2.5.1
hpack==3.0.0
html5lib==1.1
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
kombu==4.6.11
newrelic==7.2.4.171
oauthlib==3.1.1
Pillow==9.0.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
py-vapid==1.8.2
pycountry==20.7.3
pycparser==2.20
pydyf==0.1.2
PyJWT==1.7.1
pyphen==0.12.0
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.13.0
requests==2.25.1
requests-oauthlib==1.3.0
rest-social-auth==6.0.1
sendgrid==6.0.5
shortuuid==1.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
stripe==2.57.0
tinycss2==1.1.1
urllib3==1.26.5
vine==1.3.0
wallet-py3k==0.0.4
wcwidth==0.2.5
weasyprint==54.3
webencodings==0.5.1
zopfli==0.2.1
